json 관련 변수, StoreItem 구조체 같은 모델 데이터 별도로 분리. Array<StoreItem>를 포괄하는 StoreModel 객체생성.
