+++ StoreApp/StoreApp/StoreTableViewController.swift
@@ -9,19 +9,6 @@
 import UIKit
 
 class StoreTableViewController: UITableViewController {
-    // 내부 경로 참조 방법
-    let jsonFilePath = Bundle.main.path(forResource: "main", ofType: "json")
-    lazy var mainJsonData = try? Data.init(contentsOf: URL(fileURLWithPath: jsonFilePath!, isDirectory: false))
-    lazy var mainJson = try JSONSerialization.jsonObject(with: mainJsonData!, options: [])
-    // StoreItem 구조체
-    struct StoreItem : Decodable{
-        var image : String
-        var title : String
-        var description : String
-        var s_price : String
-        var badge : [String]?   // 배열, 값이 없을수도 있음
-    }
-    var itemArray : Array<StoreItem> = []
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -42,43 +29,34 @@ class StoreTableViewController: UITableViewController {
 
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         // #warning Incomplete implementation, return the number of rows
-        
-        // json decode
-        let decoder = JSONDecoder()
-        do{
-            try itemArray = decoder.decode(Array<StoreItem>.self, from: mainJsonData!)
-            print(itemArray[0])
-        }catch let error{
-            print(error)
-        }
-        
-        return itemArray.count
+        let storeModel = StoreModel()
+        let storeItemArray = storeModel.decodeJsonToStoreItemArray()
+        return storeItemArray.count
     }
 
-    
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: "StoreItemCell", for: indexPath) as! StoreItemCell
-        // Configure the cell...
-        cell.itemTitleLabel.text = itemArray[indexPath.row].title
-        cell.itemDescriptionLabel.text = itemArray[indexPath.row].description
-        cell.itemPriceLabel.text = itemArray[indexPath.row].s_price
-        cell.itemSalePriceLabel.text = itemArray[indexPath.row].s_price
-        // badge 개수에 따라 ui처리
-        if(itemArray[indexPath.row].badge != nil){
-            // badge 1개 이상
-            cell.itemBadgeLabel1.text = itemArray[indexPath.row].badge?[0]
-            if((itemArray[indexPath.row].badge?.count)! > 1){
-                cell.itemBadgeLabel2.text = itemArray[indexPath.row].badge?[1]
-            }
-            else{
-                cell.itemBadgeLabel2.isHidden = true
-            }
-        }
-        else{
-            // badge 0개. 안보이게 처리.
-            cell.itemBadgeLabel1.isHidden = true
-            cell.itemBadgeLabel2.isHidden = true
-        }
+//        // Configure the cell...
+//        cell.itemTitleLabel.text = itemArray[indexPath.row].title
+//        cell.itemDescriptionLabel.text = itemArray[indexPath.row].description
+//        cell.itemPriceLabel.text = itemArray[indexPath.row].s_price
+//        cell.itemSalePriceLabel.text = itemArray[indexPath.row].s_price
+//        // badge 개수에 따라 ui처리
+//        if(itemArray[indexPath.row].badge != nil){
+//            // badge 1개 이상
+//            cell.itemBadgeLabel1.text = itemArray[indexPath.row].badge?[0]
+//            if((itemArray[indexPath.row].badge?.count)! > 1){
+//                cell.itemBadgeLabel2.text = itemArray[indexPath.row].badge?[1]
+//            }
+//            else{
+//                cell.itemBadgeLabel2.isHidden = true
+//            }
+//        }
+//        else{
+//            // badge 0개. 안보이게 처리.
+//            cell.itemBadgeLabel1.isHidden = true
+//            cell.itemBadgeLabel2.isHidden = true
+//        }
         return cell
     }
